splits vault-code into multiple smaller files
