+++ lib/ansible/parsing/yaml/objects.py
@@ -23,6 +23,7 @@ import sys as _sys
 
 from collections.abc import Sequence
 
+import ansible.errors
 from ansible.module_utils.six import text_type
 from ansible.module_utils.common.text.converters import to_bytes, to_text, to_native
 
@@ -79,7 +80,7 @@ class AnsibleVaultEncryptedUnicode(Sequence, AnsibleBaseYAMLObject):
     @classmethod
     def from_plaintext(cls, seq, vault, secret):
         if not vault:
-            raise vault.AnsibleVaultError('Error creating AnsibleVaultEncryptedUnicode, invalid vault (%s) provided' % vault)
+            raise ansible.errors.AnsibleVaultError('Error creating AnsibleVaultEncryptedUnicode, invalid vault (%s) provided' % vault)
 
         ciphertext = vault.encrypt(seq, secret)
         avu = cls(ciphertext)
